Separate DockerManager and Hypervisor classes (#3219)
* Refactor golem.docker.manager module (#3223)

* Split Command and Hypervisor classes from golem.docker.manager

* Split Command and Hypervisor tests

* Review: DockerForMac: move out _pid and _pipe

* Review: exception types

* Review: rename docker_vm to vm_name

* Review: add 'name = name or self._vm_name' to Hypervisor.remove
